decls: PS2 float arg order / pointer args, drop 13 asm aliases, 17 COMPILER-DIFF tags

+++ src/em26/em26.cpp
@@ -383,7 +383,7 @@ static void em26_R1_Wait(cEm26* em)
         w->estTimer--;
     } else {
         w->estTimer = Rnd() % 30 + 90;
-        EstSet((int) em, -1, 0, 0, 0x1E, 3, 0, 0, (u32) em, 0);
+        EstSet(em, -1, 0, 0, 0x1E, 3, 0, 0, em, 0);
     }
     if (w->dmgTotal > 500) {
         cModel* p = em->getPartsPtr(4);
@@ -411,9 +411,9 @@ static void em26_R1_Atk(cEm26* em)
             mode = 0x41;
         }
         if (ang < 0.0f) {
-            MotionSetCore(em, MOTION(em), ARC(0xE), (int) ARC(0x16), 0, mode, 0);
+            MotionSetCore(em, MOTION(em), ARC(0xE), ARC(0x16), 0, mode, 0);
         } else {
-            MotionSetCore(em, MOTION(em), ARC(0xF), (int) ARC(0x17), 0, mode, 0);
+            MotionSetCore(em, MOTION(em), ARC(0xF), ARC(0x17), 0, mode, 0);
         }
         w->atkHit = 0;
         em->r_no_2++;
@@ -431,7 +431,7 @@ static void em26_R1_Atk(cEm26* em)
         w->estTimer--;
     } else {
         w->estTimer = Rnd() % 20 + 10;
-        EstSet((int) em, -1, 0, 0, 0x1E, 3, 0, 0, (u32) em, 0);
+        EstSet(em, -1, 0, 0, 0x1E, 3, 0, 0, em, 0);
     }
 }
 
@@ -537,11 +537,11 @@ static void em26_R1_Die_Normal(cEm26* em)
         if (w->flags & 0x10) {
             mode = 0x41;
         }
-        MotionSetCore(em, MOTION(em), ARC(0xD), (int) seq, 0, mode, 0);
+        MotionSetCore(em, MOTION(em), ARC(0xD), seq, 0, mode, 0);
         em->atari.m_flag &= ~0x200;
         SndStop(w->sndId, 0);
         w->sndId = SndCall(8, 8, &em->pos, em->id, 0, em);
-        EstSet((int) em, -1, 0, 0, 0x1E, 2, 0, 0, (u32) em, 0);
+        EstSet(em, -1, 0, 0, 0x1E, 2, 0, 0, em, 0);
         em->r_no_2++;
     }
     case 1:
