+++ src/game/ctrl14.cpp
@@ -100,7 +100,7 @@ void cCtrl14::move()
     if (w->fireDelay) {
         w->fireDelay--;
         if (w->fireDelay == 0) {
-            EstSet((int) w->obj[1], -1, NULL, NULL, 1, 2, 1, w->espKind, (u32) w->obj[1], NULL);
+            EstSet(w->obj[1], -1, NULL, NULL, 1, 2, 1, w->espKind, w->obj[1], NULL);
             SndCall(6, 2, &w->obj[1]->pos, 0, 0, w->obj[1]);
             w->fireTimer = 60;
         }
@@ -109,7 +109,7 @@ void cCtrl14::move()
         w->fireTimer--;
         if (w->fireTimer == 0) {
             EffectEspgenDelete(0, w->espKind, w->obj[1]);
-            EstSet((int) w->obj[1], -1, NULL, NULL, 1, 4, 1, 0, (u32) w->obj[1], NULL);
+            EstSet(w->obj[1], -1, NULL, NULL, 1, 4, 1, 0, w->obj[1], NULL);
         }
     }
     pos = w->obj[1]->getPartsPtr(0)->world;
@@ -333,10 +333,10 @@ void cCtrl14::setFire()
         if (w->fireTimer) {
             w->fireTimer = 0;
             EffectEspgenDelete(0, w->espKind, o);
-            EstSet((int) w->obj[1], -1, NULL, NULL, 1, 4, 1, 0, (u32) w->obj[1], NULL);
+            EstSet(w->obj[1], -1, NULL, NULL, 1, 4, 1, 0, w->obj[1], NULL);
         }
         w->fireDelay = 30;
-        EstSet((int) w->obj[1], -1, NULL, NULL, 1, 3, 1, w->espKind, (u32) w->obj[1], NULL);
+        EstSet(w->obj[1], -1, NULL, NULL, 1, 3, 1, w->espKind, w->obj[1], NULL);
         SndCall(6, 1, &w->obj[1]->pos, 0, 0, w->obj[1]);
     }
 }
